Add scripts for copying built astra library into Conda env
conda_copy* scripts are made for automatic copy of astra c++ lib into the conda environment and for building the python interface.
Those scripts are written by looking at the build_*.bat scripts.

The library should be built using MSVC as written in Astra's documentation.

There are 2 versions of the debug scripts in order to try a working configuration, but none of them work so far.
conda_copy.bat works fine.